requirements.txt
django
djangorestframework
djangorestframework_simplejwt
httpie
djongo
sqlparse==0.2.4